Create intel-types crate
Also moves Masked type and XSave type from sgx-types
to intel-types.
